+++ src/Application/Query/Catalog/SheetPreviewViewQueryHandler.php
@@ -15,6 +15,7 @@ use Proximum\Vimeet\Application\Components\Sheet\Preview\Preview;
 use Proximum\Vimeet\Application\Components\Sheet\SheetInfoGuesser;
 use Proximum\Vimeet\Application\View\Sheet\Catalog\CatalogSheetPreviewView;
 use Proximum\Vimeet\Domain\KeyDates\Checker\MeetingPublishedAccessChecker;
+use Proximum\Vimeet\Domain\Model\Meeting\Request;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Repository\Meeting\RequestRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\RuleRepositoryInterface;
@@ -138,7 +139,8 @@ class SheetPreviewViewQueryHandler
             null !== $meetingRequest ? $meetingRequest->hasMessage() : false,
             $catalogSheetPreviewViewQuery->isSeenByCurrentUser,
             $catalogSheetPreviewViewQuery->isMobileValidationRequired,
-            $validatePhoneLink ?? null
+            $validatePhoneLink ?? null,
+            $catalogSheetPreviewViewQuery->isPriority
         );
     }