+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20190620124431.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\DBAL\Schema\Schema;
+
+/**
+ * Add priority in Meeting/Request
+ */
+class Version20190620124431 extends AbstractMigration
+{
+    public function up(Schema $schema)
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE meeting_request ADD to_priority TINYINT(1) DEFAULT \'0\' NOT NULL, ADD from_priority TINYINT(1) DEFAULT \'0\' NOT NULL;');
+    }
+
+    public function down(Schema $schema)
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE meeting_request DROP to_priority, DROP from_priority');
+    }
+}