+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VIMEET-2009 - Google and Linkedin login enabling on one event
 - VIMEET-1081 - Priority meeting request number by type
 - VIMEET-1648 - Terms of sale links
+- VIMEET-1082 - Prioritize meeting request
 - VIMEET-2020 - Get uploaded sheets files from a public route
 ### Fixed
 - MV-184 - Fix rooming list assign dates